Bug 1306709 - Move runnable jobs api logic into the runnable_jobs module
Until now, the only way to get runnable jobs information is by
querying the Treeherder runnable jobs api.

After this change, Treeherder modules won't need to call the API
but import the same function the API calls (list_runnable_jobs()).
